refactor(mespapiers): Move the logic of the OCR process into its modal
