+++ packages/cozy-mespapiers-lib/src/components/ModelSteps/ScanResult/OcrProcessingDialog.jsx
@@ -1,13 +1,50 @@
-import React from 'react'
+import PropTypes from 'prop-types'
+import React, { useEffect } from 'react'
 
+import { useWebviewIntent } from 'cozy-intent'
 import { Dialog } from 'cozy-ui/transpiled/react/CozyDialogs'
 import Empty from 'cozy-ui/transpiled/react/Empty'
 import { useI18n } from 'cozy-ui/transpiled/react/providers/I18n'
 
 import OcrProcessingIcon from '../../../assets/icons/OcrProcessing.svg'
+import { useFormData } from '../../Hooks/useFormData'
+import { useStepperDialog } from '../../Hooks/useStepperDialog'
+import {
+  getAttributesFromOcr,
+  getFormDataFilesForOcr,
+  makeMetadataFromOcr
+} from '../helpers'
 
-const OcrProcessingDialog = ({ onBack }) => {
+const OcrProcessingDialog = ({ onBack, rotatedFile }) => {
   const { t } = useI18n()
+  const { setFormData, formData } = useFormData()
+  const webviewIntent = useWebviewIntent()
+  const { currentDefinition, nextStep } = useStepperDialog()
+  const { ocrAttributes } = currentDefinition
+
+  useEffect(() => {
+    const init = async () => {
+      const files = getFormDataFilesForOcr(formData, rotatedFile)
+      const attributesFound = await getAttributesFromOcr({
+        files,
+        ocrAttributes,
+        webviewIntent
+      })
+
+      const metadataFromOcr = makeMetadataFromOcr(attributesFound)
+      setFormData(prev => ({
+        ...prev,
+        metadata: {
+          ...prev.metadata,
+          ...metadataFromOcr
+        }
+      }))
+
+      nextStep()
+    }
+    init()
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [])
 
   return (
     <Dialog
@@ -25,6 +62,11 @@ const OcrProcessingDialog = ({ onBack }) => {
       }
     />
   )
+}
+
+OcrProcessingDialog.propTypes = {
+  onBack: PropTypes.func.isRequired,
+  rotatedFile: PropTypes.object
 }
 
 export default OcrProcessingDialog
